完善优化统计界面

+++ src/main/java/com/netresmanager/model/StatsSummary.java
@@ -4,7 +4,7 @@ package com.netresmanager.model;
  * Summary statistics for a project or across all projects.
  */
 public class StatsSummary {
-    public int totalProcessed;
+    public long totalProcessed;
     public long totalExported;
     public long totalRecycled;
     public int uniqueFileTypes;
